Move content from `WireUI.WireConversationUI` to `WireMessaging.WireMessagingUI`
