+++ WireMessaging/.swiftpm/xcode/xcshareddata/xcschemes/WireMessagingAll.xcscheme
@@ -58,7 +58,7 @@
       shouldUseLaunchSchemeArgsEnv = "YES">
       <TestPlans>
          <TestPlanReference
-            reference = "container:WireMessaging/Tests/TestPlans/AllTests.xctestplan"
+            reference = "container:Tests/TestPlans/AllTests.xctestplan"
             default = "YES">
          </TestPlanReference>
       </TestPlans>
