+++ WireMessaging/Sources/WireMessagingUI/Conversation/ConversationCell/CellTypes/ConversationCellModel+configureCell.swift
@@ -16,8 +16,20 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-import Foundation
+public import UIKit
 
-public let SnapshotTestReferenceImageDirectory = URL(fileURLWithPath: #filePath)
-    .deletingLastPathComponent()
-    .path
+public extension ConversationCellModel {
+
+    @MainActor
+    func configureCell(_ cell: UITableViewCell) {
+        switch self {
+
+        case let .timeDivider(timeDivider):
+            guard let cell = cell as? ConversationCell<TimeDividerModel> else { break }
+            return cell.model = timeDivider
+        }
+
+        assertionFailure("unexpected cell: \(cell)")
+    }
+
+}
