+++ WireMessaging/Sources/WireMessagingUI/Conversation/ConversationTitleView/ConversationTitleSource.swift
@@ -16,7 +16,7 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-import WireAccountImageUI
+public import WireAccountImageUI
 
 public struct ConversationTitleSource {
     public let accountImageSource: AccountImageSource?
