+++ WireMessaging/Sources/WireMessagingUI/Resources/Images.xcassets/CertificateValid.imageset/Contents.json
