+++ WireMessaging/Sources/WireMessagingUI/Resources/Images.xcassets/Contents.json
