+++ WireMessaging/Sources/WireMessagingUI/Resources/Images.xcassets/Dropdown.imageset/Contents.json
