+++ WireMessaging/Sources/WireMessagingUI/Resources/Images.xcassets/LegalHold.imageset/Contents.json
