+++ WireMessaging/Sources/WireMessagingUI/Resources/Images.xcassets/Verified.imageset/Contents.json
