+++ WireMessaging/Tests/WireMessagingUITests/Resources/ReferenceImages/ConversationCellTimeDividerSnapshotTests/testUIFontContentSizeCategories.accessibilityExtraExtraExtraLarge.png
