+++ WireMessaging/Tests/WireMessagingUITests/Resources/ReferenceImages/ConversationCellTimeDividerSnapshotTests/testUIFontContentSizeCategories.small.png
