+++ WireMessaging/Tests/WireMessagingUITests/Resources/ReferenceImages/ConversationTitleViewSnapshotTests/testImageAvatarForOneOnOnesWithNoSubtitle.1.png
