+++ WireMessaging/Tests/WireMessagingUITests/Resources/ReferenceImages/ConversationTitleViewSnapshotTests/testInitialsAvatarForOneOnOnesWithSubtitle.1.png
