+++ WireMessaging/Tests/WireMessagingUITests/Resources/ReferenceImages/ConversationTitleViewSnapshotTests/testInitialsAvatarForOneOnOnesWithSubtitle._dark.png
