+++ WireMessaging/Tests/WireMessagingUITests/Resources/ReferenceImages/ConversationTitleViewSnapshotTests/testInitialsAvatarForOneOnOnesWithSubtitle.dark.png
