+++ WireMessaging/Tests/WireMessagingUITests/Resources/ReferenceImages/ConversationTitleViewSnapshotTests/testNoAvatarForGroupConversation.1.png
