+++ WireMessaging/Tests/WireMessagingUITests/Resources/ReferenceImages/ConversationTitleViewSnapshotTests/testWithAllContent.1.png
