--- WireUI/Sources/WireConversationUI/WireConversationUI.docc/Documentation.md
@@ -1,15 +1,0 @@
-# ``WireConversationUI``
-
-Code for presenting the content (messages) of conversations.
-
-## Overview
-
-Encapsulates all the code and functionality related to displaying and managing
-the messages of conversations within the app.
-
-## Topics
-
-### UIDiffableDataSource
-
-Check out ``ConversationCellsPreview`` in order to see how ``ConversationCellModel``
-can be used to populate a table view with little effort.
