+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConfigurationMessageCell/Components/BurstTimestampTableViewCell.swift
@@ -17,7 +17,7 @@
 //
 
 import UIKit
-import WireConversationUI
+import WireMessagingUI
 import WireFoundation
 
 final class BurstTimestampSenderMessageCellDescription: ConversationMessageCellDescription {
