+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConfigurationMessageCell/ConversationMessageCell.swift
@@ -17,7 +17,7 @@
 //
 
 import UIKit
-import WireConversationUI
+import WireMessagingUI
 import WireDataModel
 import WireUtilities
