+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/ConversationViewController.swift
@@ -19,7 +19,6 @@
 import UIKit
 import WireCommonComponents
 import WireMessagingUI
-import WireConversationUI
 import WireDesign
 import WireLogging
 import WireMainNavigationUI
@@ -87,7 +86,7 @@ final class ConversationViewController: UIViewController {
     let invisibleInputAccessoryView: InvisibleInputAccessoryView = .init()
     private let mediaBarViewController: MediaBarViewController
 
-    private let titleView: WireConversationUI.ConversationTitleView
+    private let titleView: WireMessagingUI.ConversationTitleView
 
     let userSession: UserSession
 
@@ -164,7 +163,7 @@ final class ConversationViewController: UIViewController {
 
         self.mediaBarViewController = MediaBarViewController(mediaPlaybackManager: mediaPlaybackManager)
 
-        self.titleView = WireConversationUI.ConversationTitleView(
+        self.titleView = WireMessagingUI.ConversationTitleView(
             source: ConversationTitleSource(
                 accountImageSource: nil,
                 title: conversation.displayNameWithFallback,
